aexpect: Use internal state to avoid bug closing Spawn objects multiple times
Whenever we .close() a Spawn() object, we're going to
clean up its base directory, removing all traces of that
object on the filesystem. Some times, though, we might
call .close() multiple times, creating a situation when
we try to recover files that were already deleted. Fix
that by managing whether objects were closed by using
a new attribute, .closed, that is set to True while it's
not closed for the first time. After the first close, all
subsequent calls to .close() will be no-ops, avoiding the
bug and keeping the temporary files created by aexpect
non predictable and secure.

+++ aexpect/client.py
@@ -110,6 +110,7 @@ class Spawn(object):
         """
         self.a_id = a_id or data_factory.generate_random_string(8)
         self.log_file = None
+        self.closed = False
 
         base_dir = os.path.join(BASE_DIR, 'aexpect_%s' % self.a_id)
 
@@ -327,18 +328,20 @@ class Spawn(object):
 
         :param sig: The signal to send the process when attempting to kill it.
         """
-        self.kill(sig=sig)
-        # Wait for the server to exit
-        wait_for_lock(self.lock_server_running_filename)
-        # Call all cleanup routines
-        for hook in self.close_hooks:
-            hook(self)
-        # Close reader file descriptors
-        self._close_reader_fds()
-        self.reader_fds = {}
-        # Remove all used files
-        if 'AEXPECT_DEBUG' not in os.environ:
-            shutil.rmtree(os.path.join(BASE_DIR, 'aexpect_%s' % self.a_id))
+        if not self.closed:
+            self.kill(sig=sig)
+            # Wait for the server to exit
+            wait_for_lock(self.lock_server_running_filename)
+            # Call all cleanup routines
+            for hook in self.close_hooks:
+                hook(self)
+            # Close reader file descriptors
+            self._close_reader_fds()
+            self.reader_fds = {}
+            # Remove all used files
+            if 'AEXPECT_DEBUG' not in os.environ:
+                shutil.rmtree(os.path.join(BASE_DIR, 'aexpect_%s' % self.a_id))
+            self.closed = True
 
     def set_linesep(self, linesep):
         """
